chore: removing unnecessary component
